GH-15: ListTile in eigene Datei
