+++ lib/views/next_routes.dart
@@ -7,6 +7,7 @@ import 'dart:async';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
 import 'package:gethome/services/current_location_service.dart';
 import 'package:geolocator/geolocator.dart';
+import 'package:gethome/views/list_tile_of_route.dart';
 
 
 class RouteSample extends StatefulWidget {
@@ -123,7 +124,7 @@ class RoutesScreen extends State<RouteSample> {
               child: Text(_errorMessage),
             )
             :
-            MyListTile(route: _nextRoutes![0])
+            RouteListTile(route: _nextRoutes![0])
           ),
           const Divider(),
 
@@ -133,7 +134,7 @@ class RoutesScreen extends State<RouteSample> {
               child: Text(_errorMessage),
             )
             :
-            MyListTile(route: _nextRoutes![1])
+            RouteListTile(route: _nextRoutes![1])
           ),
           const Divider(),
 
@@ -143,7 +144,7 @@ class RoutesScreen extends State<RouteSample> {
               child: Text(_errorMessage),
             )
             :
-            MyListTile(route: _nextRoutes![2])
+            RouteListTile(route: _nextRoutes![2])
           ),
           const Divider(),
         ],
@@ -158,93 +159,4 @@ class RoutesScreen extends State<RouteSample> {
     ),
     );
   }
-}
-
-/// Class of a ListTile that takes a GetHomeRoute and returns a complete designed
-/// ListTile that displays the route's data.
-class MyListTile extends StatelessWidget {
-  // route to be displayed
-  final GetHomeRoute route;
-
-  MyListTile({required this.route});
-
-  @override
-  Widget build(BuildContext context) {
-    return ListTile(
-      contentPadding: const EdgeInsets.symmetric(horizontal: 30.0),
-      leading: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: [
-
-          // rectangle with the name of the first line inside
-          Container(
-            width: 40,
-            height: 30,
-            alignment: Alignment.center,
-            decoration: BoxDecoration(
-              color: (route.color == null ? Colors.blue : Color(route.color!)), // Farbe des Rechtecks
-              borderRadius: BorderRadius.circular(5), // Runden des Rechtecks
-            ),
-            child: Text(
-              route.firstLine.toString(),
-              style: const TextStyle(
-                color: Colors.white,
-                fontWeight: FontWeight.bold,
-                fontSize: 18,
-              ),
-            ),
-          ),
-
-        ],
-      ),
-      title: Row(
-        children: [
-
-          // displaying the number of changes
-          Icon(Icons.arrow_forward_ios,
-          size: 14),
-          Text(
-            '+${(route.changes == null || route.changes == 0 ? 1 : route.changes!) - 1}',
-            style: TextStyle(fontSize: 14), // Kleinerer Text für changes
-          ),
-
-          // empty space for separation
-          SizedBox(width: 20),
-
-          // displaying the departure time of the first line
-          Text(
-            extractTime(route.departureTime!),
-            style: TextStyle(fontSize: 25),
-          ),
-        ],
-      ),
-
-      trailing: Row(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-
-          // displaying the walking time and a walking person icon
-          Icon(Icons.directions_walk,),
-          SizedBox(width: 5),
-          Text(
-            '${(route.walkingTimeMinutes == null ? 0 : route.walkingTimeMinutes!) ~/ 60} min',
-            style: TextStyle(fontSize: 20),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-
-/// Additional method to convert a DateTime to a String that consists only of hours
-/// and minutes.
-String extractTime(DateTime dateTime) {
-  // Konvertiere DateTime in einen String
-  String dateTimeString = dateTime.toLocal().toString();
-  
-  // Extrahiere nur die Zeit
-  String timeString = dateTimeString.split(' ')[1];
-
-  return '${timeString.split(':')[0]}:${timeString.split(':')[1]}';
 }
